Moving test to it's own file so we can test more css
